added OLCC and drinking

+++ a3/Wallet.java
@@ -9,13 +9,13 @@
 public class Wallet
 {
     
-    private int money;
+    private double money;
 
     /**
      * Constructor for objects of class Wallet
      * @param money how much money is in the Wallet
      */
-    public Wallet(int money) {
+    public Wallet(double money) {
         this.money = money;
     }
     
@@ -24,7 +24,7 @@ public class Wallet
      * 
      * @return money - the current amount of money in the wallet
      */
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
@@ -34,7 +34,7 @@ public class Wallet
      * @param  cash  the amount of money you want to add to your wallet
      * @return  void
      */
-    public void addMoney(int cash) {
+    public void addMoney(double cash) {
         money += cash;
     }
     
@@ -45,7 +45,7 @@ public class Wallet
      * @return  void
      * @throws XX if cash is > than money in the wallet
      */
-    public void removeMoney(int cash) {
+    public void removeMoney(double cash) {
         
         if (cash > money) {
             throw new IllegalArgumentException ("You cannot remove more money than you have in your wallet.");
